requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.169.0
google-auth>=2.40.0
google-auth-httplib2>=0.2.0
